karatoken-integrated/ai-services/scoring-service/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
numpy==1.24.2
librosa==0.10.0
soundfile==0.12.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic==1.10.7
python-socketio==5.8.0
openai-whisper==20230314
pydub==0.25.1
